+++ CHANGES.rst
@@ -7,7 +7,11 @@ Changelog
 - Display `content_category` title at the top of `@@categorized-childs-infos`
   tooltipster view.
   [gbastien]
-- Use `natsort.realsorted` to sort categorized elements.
+- Use `natsort.realsorted` to sort categorized elements on their title,
+  elements are sorted regardless of uppercase or lowercase title.
+  [gbastien]
+- Moved `context._p_changed = True` to the `utils.sort_categorized_elements`
+  method so we are sure that calling it will correctly manage `_p_changed`.
   [gbastien]
 
 0.33 (2018-08-03)